+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 **Features:**
 - Added #1171: Show start writing message when no posts.
 - Added #1173: Discord link to dashboard.
+- Added #1185: Bookmarking for Reader.
 
 **Bug Fixes:**
 - Fixed #1155: Fix top margin on main wrapper.